basic zoom support

+++ src/map.c
@@ -3,8 +3,8 @@
 //50 => 250
 //0 => 50
 //0 => 0
-
-double map (int n, int f1, int t1, int f2, int t2)
+#include <math.h>
+double map (double n, double f1, double t1, double f2, double t2)
 {
 	double def1 = t1 - f1;
 	double def2 = t2 - f2;
